// ui-tui/src/app/interfaces.ts
export interface UiState {
  bgTasks: Set<string>
  busy: boolean
  compact: boolean
  detailsMode: DetailsMode
  info: null | SessionInfo
  inlineDiffs: boolean
  mouseTracking: boolean
  showCost: boolean
  showReasoning: boolean
  sid: null | string
  status: string
  statusBar: boolean
  streaming: boolean
  theme: Theme
  usage: Usage
}